chore: rename predicterrorstate
