--- app/components/UI/Predict/components/PredictErrorState/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PredictErrorState';
