+++ app/components/UI/Predict/components/PredictOffline/PredictOffline.styles.ts
