+++ app/components/UI/Predict/components/PredictOffline/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PredictOffline';
